csync: Move the locale<->utf8 conversion to a different file
It's a C++ file so we will be able to use Qt from it
