+++ csync/src/std/c_string.c
@@ -167,62 +167,3 @@ void c_strlist_destroy(c_strlist_t *strlist) {
   SAFE_FREE(strlist->vector);
   SAFE_FREE(strlist);
 }
-
-/* Convert a wide multibyte String to UTF8 */
-char* c_utf8_from_locale(const mbchar_t *wstr)
-{
-  char *dst = NULL;
-#ifdef _WIN32
-  char *mdst = NULL;
-  int size_needed;
-  size_t len;
-#endif
-
-  if (wstr == NULL) {
-    return NULL;
-  }
-
-#ifdef _WIN32
-  len = wcslen(wstr);
-  /* Call once to get the required size. */
-  size_needed = WideCharToMultiByte(CP_UTF8, 0, wstr, len, NULL, 0, NULL, NULL);
-  if (size_needed > 0) {
-    mdst = c_malloc(size_needed + 1);
-
-    memset(mdst, 0, size_needed + 1);
-    WideCharToMultiByte(CP_UTF8, 0, wstr, len, mdst, size_needed, NULL, NULL);
-    dst = mdst;
-  }
-#else
-  dst = c_strdup(wstr);
-#endif
-  return dst;
-}
-
-/* Convert a an UTF8 string to multibyte */
-mbchar_t* c_utf8_string_to_locale(const char *str)
-{
-    mbchar_t *dst = NULL;
-#ifdef _WIN32
-    size_t len;
-    int size_needed;
-#endif
-
-    if (str == NULL ) {
-        return NULL;
-    }
-
-#ifdef _WIN32
-    len = strlen(str);
-    size_needed = MultiByteToWideChar(CP_UTF8, 0, str, len, NULL, 0);
-    if (size_needed > 0) {
-        int size_char = (size_needed + 1) * sizeof(mbchar_t);
-        dst = c_malloc(size_char);
-        memset((void*)dst, 0, size_char);
-        MultiByteToWideChar(CP_UTF8, 0, str, -1, dst, size_needed);
-    }
-#else
-    dst = c_strdup(str);
-#endif
-    return dst;
-}
